fix mentee delete and add swal to delete

+++ src/Components/table/optionAction.js
@@ -4,12 +4,39 @@ import {setMenteeEditId} from "../../store/reducer/mentee.js";
 import {setMentorEditId} from "../../store/reducer/mentor.js";
 import {setUnivEditId} from "../../store/reducer/univ.js";
 import {setTugasEditId} from "../../store/reducer/tugas.js";
+import Swal from "sweetalert2";
 
 
 
 export const mentee={
-    delete:(id)=>{
+    delete:(id, handleRefresh)=>{
         console.log("delete ",id)
+        Swal.fire({
+            title:"Apakah anda yakin?",
+            text:"Setelah menghapus data tidak akan ditampilkan lagi!",
+            icon:"warning",
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText:'Iya, hapus!',
+            cancelButtonText:"Tidak"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`/mentee/${id}`)
+                    .then(res=>{
+                        console.log(res.data.msg)
+                        handleRefresh()
+                    })
+                    .catch(err=>{
+                        Swal.fire({
+                            title: "Terjadi Kesalahan",
+                            text: "Periksa kembali Koneksi",
+                            icon: "error"
+                        })
+                        console.log(err)
+                    })
+            }
+        })
     },
     edit:(id, dispatch, openInputModal)=>{
         console.log("edit ",id)
@@ -39,15 +66,32 @@ export const admin={
 export const mentor={
     delete:(id, handleRefresh)=>{
         console.log("delete ",id)
-        axios.delete(`/mentor/${id}`)
-            .then(res=>{
-                console.log(res.data.msg)
-                handleRefresh()
-            })
-            .catch(err=>{
-                console.log(err)
-            })
-
+        Swal.fire({
+            title:"Apakah anda yakin?",
+            text:"Setelah menghapus data tidak akan ditampilkan lagi!",
+            icon:"warning",
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText:'Iya, hapus!',
+            cancelButtonText:"Tidak"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`/mentor/${id}`)
+                    .then(res=>{
+                        console.log(res.data.msg)
+                        handleRefresh()
+                    })
+                    .catch(err=>{
+                        Swal.fire({
+                            title: "Terjadi Kesalahan",
+                            text: "Periksa Koneksi dan pastikan mentee tidak terasosiasi",
+                            icon: "error"
+                        })
+                        console.log(err)
+                    })
+            }
+        })
     },
     edit:(id, dispatch, openInputModal)=>{
         console.log("edit ",id)
@@ -58,15 +102,32 @@ export const mentor={
 export const universitas={
     delete:(id, handleRefresh)=>{
         console.log("delete ",id)
-        axios.delete(`/universitas/${id}`)
-            .then(res=>{
-                console.log(res.data.msg)
-                handleRefresh()
-            })
-            .catch(err=>{
-                console.log(err)
-            })
-
+        Swal.fire({
+            title:"Apakah anda yakin?",
+            text:"Setelah menghapus data tidak akan ditampilkan lagi!",
+            icon:"warning",
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText:'Iya, hapus!',
+            cancelButtonText:"Tidak"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`/universitas/${id}`)
+                    .then(res=>{
+                        console.log(res.data.msg)
+                        handleRefresh()
+                    })
+                    .catch(err=>{
+                        Swal.fire({
+                            title: "Terjadi Kesalahan",
+                            text: "Periksa Koneksi dan pastikan mentee tidak terasosiasi",
+                            icon: "error"
+                        })
+                        console.log(err)
+                    })
+            }
+        })
     },
     edit:(id, dispatch, openInputModal)=>{
         console.log("edit ",id)
